Fix all clippy pedantic/nursery warnings; add justfile and clippy rules to agent.md
- use_self: replace TypeName:: with Self:: in all Display impls (types.rs)
- if_not_else: invert torrent check condition to positive form (core/mod.rs)
- redundant_pub_crate: pub(crate) -> pub in private shell modules (docker.rs, mam.rs)
- map_unwrap_or: .map(..).unwrap_or(false) -> .is_some_and(..) (docker.rs, mam.rs)
- needless_pass_by_value: watch_dir String -> &str in spawn_file_watcher
- clone_on_copy: val.clone() -> *val for Copy type (docker.rs)
- map_unwrap_or_else: .map(..).unwrap_or_else(..) -> .map_or_else(..) (mod.rs)
- collapsible_if: flatten nested if/if-let in authenticate (qbit.rs)
- unnecessary_cast: remove redundant as u64 in statvfs calculation (monitors.rs)
- cast_precision_loss: allow with reason comment for bytes->f64 (monitors.rs)
- too_many_lines: allow with reason comment on dispatch_one (mod.rs)
- doc_markdown: add backticks around ReadPortFiles in doc comment (events.rs)

+++ src/shell/monitors.rs
@@ -6,6 +6,9 @@ use uom::si::information::byte;
 /// so the Core never triggers a false low-space alert.
 pub fn check_disk_space(path: &str) -> Information {
     match available_bytes(path) {
+        #[allow(clippy::cast_precision_loss)]
+        // f64 mantissa is 52 bits (~8 PiB exact range). Practical disk sizes
+        // are well under this, so the precision loss is negligible for alerting.
         Ok(bytes) => Information::new::<byte>(bytes as f64),
         Err(e) => {
             warn!("Failed to check disk space at {path}: {e}");
@@ -25,7 +28,7 @@ fn available_bytes(path: &str) -> anyhow::Result<u64> {
         anyhow::bail!("statvfs failed: {}", std::io::Error::last_os_error());
     }
     let stat = unsafe { stat.assume_init() };
-    Ok(stat.f_bavail as u64 * stat.f_frsize as u64)
+    Ok(stat.f_bavail * stat.f_frsize)
 }
 
 #[cfg(test)]
